Move xkeyboard into the keyboard module
