+++ rust/agama-locale-data/src/keyboard.rs
@@ -1,3 +1,5 @@
 pub mod xkb_config_registry;
+pub mod xkeyboard;
 
 pub use xkb_config_registry::XkbConfigRegistry;
+pub use xkeyboard::XKeyboards;
